Deleted Subfolder JS/CSS/Fonts
Every reference to the js, css, and font libs refer to the folders in
the root of the site. There is no reason to have multiple instances of
these static libs and will make them hard to replace if there’s
theoretically ever any initiative to do that.

This could also be a performance gain on browsers that cache these
sorts of things since throughout the site the same files are constantly
being referenced.

I also updated the HTML template to reflect these changes so no one
will have to think about this when they are using it.